app/requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pandas==2.2.2
streamlit==1.38.0
altair==5.4.1
folium==0.17.0
branca==0.7.2
shapely==2.0.6
streamlit_folium==0.22.1